requirements.txt
langchain
langchain-openai
langchain-community
atlassian-python-api
python-dotenv
chromadb
tiktoken
ibm-watsonx-orchestrate
